+++ orphecoresdk/src/main/java/io/orphe/orphecoresdk/OrpheInsoleValue.java
@@ -88,9 +88,6 @@ public class OrpheInsoleValue {
       this.quatX = 0.0;
       this.quatY = 0.0;
       this.quatZ = 0.0;
-      this.eulerYaw = 0.0;
-      this.eulerPitch = 0.0;
-      this.eulerRoll = 0.0;
       this.receivedAt = receivedAt;
     }
 
@@ -140,8 +137,9 @@ public class OrpheInsoleValue {
                           @NonNull final long receivedAt
 
                      ){
-      final OrpheEulerAngles euler =
-              OrpheEulerAngles.fromQuaternion(quatW, quatX, quatY, quatZ);
+//      Euler角の計算は一時停止しています。
+//      final OrpheEulerAngles euler =
+//              OrpheEulerAngles.fromQuaternion(quatW, quatX, quatY, quatZ);
       this.sidePosition = sidePosition;
       this.serialNumber = serialNumber;
       this.dataPosition = dataPosition;
@@ -163,9 +161,6 @@ public class OrpheInsoleValue {
       this.quatX = quatX;
       this.quatY = quatY;
       this.quatZ = quatZ;
-      this.eulerYaw = euler.yaw;
-      this.eulerPitch = euler.pitch;
-      this.eulerRoll = euler.roll;
       this.receivedAt = receivedAt;
     }
 
@@ -208,13 +203,14 @@ public class OrpheInsoleValue {
         builder.append(",");
         builder.append(String.format("%.2f", gyroZ));
         builder.append(")\n");
-        builder.append("euler(yaw,pitch,roll):(");
-        builder.append(String.format("%.2f", eulerYaw));
-        builder.append(",");
-        builder.append(String.format("%.2f", eulerPitch));
-        builder.append(",");
-        builder.append(String.format("%.2f", eulerRoll));
-        builder.append(")\n");
+//        Euler角の出力は一時停止しています。
+//        builder.append("euler(yaw,pitch,roll):(");
+//        builder.append(String.format("%.2f", eulerYaw));
+//        builder.append(",");
+//        builder.append(String.format("%.2f", eulerPitch));
+//        builder.append(",");
+//        builder.append(String.format("%.2f", eulerRoll));
+//        builder.append(")\n");
         return builder.toString();
     }
 
@@ -677,19 +673,6 @@ public class OrpheInsoleValue {
      * クオータニオンZ
      */
     @NonNull public final double quatZ;
-
-    /**
-     * ヨー角（ラジアン）
-     */
-    @NonNull public final double eulerYaw;
-    /**
-     * ピッチ角（ラジアン）
-     */
-    @NonNull public final double eulerPitch;
-    /**
-     * ロール角（ラジアン）
-     */
-    @NonNull public final double eulerRoll;
 
     /**
      * SDKが値を受信した時刻（Android端末時刻・epochミリ秒）